# server/OpenAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie API",
    "description": "API for users to access accounts, profiles, and group interactions.",
    "contact": { "email": "your-email@example.com" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://movieapp-backend-actions-gxfcdkh7h6hccvfu.northeurope-01.azurewebsites.net",
      "description": "azure development server"
    },
    {
      "url": "http://localhost:{port}/",
      "description": "Local development server",
      "variables": {
        "port": {
          "description": "Port on which the server is running",
          "default": 3000
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/group/{groupId}": {
      "get": {
        "tags": ["groups"],
        "summary": "Get Group by Group ID",
        "description": "Retrieve details of a specific group by its unique ID.",
        "operationId": "getGroupDetails",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_group_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      },
      "put": {
        "tags": ["groups"],
        "summary": "Update Group Details",
        "description": "Update the details of an existing group.",
        "operationId": "updateGroupDetails",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/update_group_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group details updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/update_group_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "get_group_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Group Name"
          },
          "description": {
            "type": "string",
            "example": "Group description"
          },
          "creator_id": {
            "type": "integer",
            "example": 10
          }
        },
        "example": {
          "id": 1,
          "name": "Study Group",
          "description": "A group for students to collaborate on homework and projects.",
          "creator_id": 10
        }
      },
      "update_group_body": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The updated name of the group",
            "example": "Updated Group Name"
          },
          "description": {
            "type": "string",
            "description": "The updated description of the group",
            "example": "Updated group description"
          }
        },
        "required": ["name", "description"]
      },
      "update_group_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Group Name"
          },
          "description": {
            "type": "string",
            "example": "Group description"
          },
          "creator_id": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
    }
  }
}